refactor(App): cleanup all code
