+++ packages/hedgehog-lab/src/components/Auth/Action/AuthAction.tsx
@@ -2,7 +2,7 @@ import {Button, Link, Typography} from "@mui/material";
 import * as React from "react";
 import {useCallback} from "react";
 import {useRecoilState} from "recoil";
-import {authActionState} from "../../../pages/Auth/RAuthStates";
+import {authActionState} from "../../../states/RAuthStates";
 
 interface IBaseActionProps {
     forget?: {
