--- packages/hedgehog-lab/src/components/Base/Form/IBaseFormMethods.ts
@@ -1,6 +1,0 @@
-export interface IBaseFormMethods {
-    login: string,
-    sign: string,
-    forget: string,
-    settingAccount: string
-}
