--- packages/hedgehog-lab/src/components/Layout/TopBar/RightButton/IRightButtonProps.ts
@@ -1,7 +1,0 @@
-import React from "react";
-
-export interface IRightButtonProps {
-    href: string,
-    render: React.ReactNode,
-    tooltip: string
-}
