--- packages/hedgehog-lab/src/config/route/IRuteProps.ts
@@ -1,7 +1,0 @@
-import * as React from "react";
-
-export interface IRuteProps {
-    path: string,
-    element: React.ReactElement,
-    children?: undefined | Array<IRuteProps>,
-}
