+++ packages/hedgehog-lab/src/modals/account/accountModal.ts
@@ -1,7 +1,7 @@
 import * as yup from "yup";
-import {IAuthFormInput} from "../../pages/Auth/IAuthFormInput";
+import {IFormInput} from "../../interfaces/IFormInput";
 
-export const accountModal = (data: IAuthFormInput): void => {
+export const accountModal = (data: IFormInput): void => {
     console.log(data)
 }
